+++ pallets/pallet-configuration/src/mock.rs
@@ -58,7 +58,7 @@ pub mod runtime {
 		type RuntimeOrigin = RuntimeOrigin;
 		type RuntimeCall = RuntimeCall;
 		type Block = Block;
-		type Nonce = u32;
+		type Nonce = u64;
 		type Hash = Hash;
 		type Hashing = BlakeTwo256;
 		type AccountId = AccountId;
