+++ pallets/pallet-did-lookup/src/mock.rs
@@ -60,7 +60,7 @@ impl frame_system::Config for Test {
 	type BlockLength = ();
 	type DbWeight = ();
 	type Block = Block;
-	type Nonce = u32;
+	type Nonce = u64;
 	type RuntimeOrigin = RuntimeOrigin;
 	type RuntimeCall = RuntimeCall;
 	type Hash = Hash;
